[lp-2189] : create limitedpartner package and move classes into it
