+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/limitedpartner/dao/LimitedPartnerDao.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.model.limitedpartner.dao;
+package uk.gov.companieshouse.limitedpartnershipsapi.limitedpartner.dao;
 
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
