+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/limitedpartner/dto/LimitedPartnerSubmissionCreatedResponseDto.java
@@ -1,0 +1,4 @@
+package uk.gov.companieshouse.limitedpartnershipsapi.limitedpartner.dto;
+
+public record LimitedPartnerSubmissionCreatedResponseDto(String id) {
+}
