--- src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/limitedpartner/dto/LimitedPartnerSubmissionCreatedResponseDto.java
@@ -1,4 +1,0 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.model.limitedpartner.dto;
-
-public record LimitedPartnerSubmissionCreatedResponseDto(String id) {
-}
