+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/personwithsignificantcontrol/PersonWithSignificantControlRepository.java
@@ -1,8 +1,9 @@
 package uk.gov.companieshouse.limitedpartnershipsapi.repository;
 
-import java.util.List;
 import org.springframework.data.mongodb.repository.MongoRepository;
-import uk.gov.companieshouse.limitedpartnershipsapi.model.personwithsignificantcontrol.dao.PersonWithSignificantControlDao;
+import uk.gov.companieshouse.limitedpartnershipsapi.personwithsignificantcontrol.dao.PersonWithSignificantControlDao;
+
+import java.util.List;
 
 public interface PersonWithSignificantControlRepository extends MongoRepository<PersonWithSignificantControlDao, String> {
     List<PersonWithSignificantControlDao> findAllByTransactionIdOrderByUpdatedAtDesc(String transactionId);
