+++ src/test/java/uk/gov/companieshouse/limitedpartnershipsapi/incorporation/IncorporationControllerValidationTest.java
@@ -22,8 +22,8 @@ import uk.gov.companieshouse.limitedpartnershipsapi.config.InterceptorConfig;
 import uk.gov.companieshouse.limitedpartnershipsapi.generalpartner.GeneralPartnerRepository;
 import uk.gov.companieshouse.limitedpartnershipsapi.generalpartner.dao.GeneralPartnerDao;
 import uk.gov.companieshouse.limitedpartnershipsapi.incorporation.dao.IncorporationDao;
+import uk.gov.companieshouse.limitedpartnershipsapi.limitedpartner.LimitedPartnerRepository;
 import uk.gov.companieshouse.limitedpartnershipsapi.partnership.PartnershipRepository;
-import uk.gov.companieshouse.limitedpartnershipsapi.repository.LimitedPartnerRepository;
 import uk.gov.companieshouse.limitedpartnershipsapi.repository.PersonWithSignificantControlRepository;
 import uk.gov.companieshouse.limitedpartnershipsapi.service.CompanyService;
 import uk.gov.companieshouse.limitedpartnershipsapi.service.CostsService;
